+++ CHANGELOG.md
@@ -33,7 +33,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Fixed
 
 - Fix `Parse` in `go.opentelemetry.io/otel/baggage` to validate member value before percent-decoding. (#4755)
-- Fix `observable` not registered error during drop aggregation in `go.opentelemetry.io/otel/sdk/metric`. (#4760)
+- Fix observable not registered error when the asynchronous instrument has a drop aggregation in `go.opentelemetry.io/otel/sdk/metric`. (#4772)
 
 ## [1.21.0/0.44.0] 2023-11-16